[Breaking Changes] Refactor all code to avoid type errors in the future (#157)
* wip

* done khaenriah

* wip

* fix format

* disable test action

* fix all type lint format

* fix news

* add typing for news

* fix ci

* remove unused isInstance

* remove unused

* fix permission

* improve some stuff

* fix

* update test workflow

* fix ci

* update readme

* update contributing.md

* refactor: streamline permission alerts and role editing responses

* chore: update version to 1.12.0 and add license and author information

* chore: update contributing guidelines and improve debug logging; bump version to 1.12.0

* fix: remove unused import and update error handler type annotation

* feat: add per-file ignores for ruff and improve test function annotations

* wip

* tested

* fix lint

* fix format

+++ bot/cogs/giveaway.py
@@ -1,9 +1,8 @@
 import logging
-from datetime import datetime, timedelta, timezone
-from typing import Optional
+from datetime import UTC, datetime, timedelta
 
 import discord
-from discord import app_commands, Interaction
+from discord import Interaction, app_commands
 from discord.ext import commands, tasks
 
 from bot.bot import WarnetBot
@@ -13,7 +12,7 @@ logger = logging.getLogger(__name__)
 
 
 @commands.guild_only()
-class Giveaway(commands.GroupCog, group_name='warnet-ga'):
+class Giveaway(commands.GroupCog, group_name="warnet-ga"):
     def __init__(self, bot: WarnetBot) -> None:
         self.bot = bot
         self.db_pool = bot.get_db_pool()
@@ -23,37 +22,54 @@ class Giveaway(commands.GroupCog, group_name='warnet-ga'):
         if not self._check_blacklist_ga.is_running():
             self._check_blacklist_ga.start()
 
-    @app_commands.command(name='blacklist', description='Blacklist a user from giveaway')
-    @app_commands.describe(
-        big='Yes if the giveaway is more than Rp50.000',
-        winners='Discord IDs of the giveaway winner. Example: 1234567890,0987654321',
-        ghosts='Member IDs who do not claim the giveaway. Example: 1234567890,0987654321',
+    async def cog_command_error(self, ctx: commands.Context, error: Exception) -> None:
+        logger.exception("An unexpected error occurred in Admin cog", exc_info=error)
+        await ctx.reply(
+            "An unexpected error occurred. Please try again later.",
+            delete_after=5,
+            ephemeral=True,
+        )
+
+    @app_commands.command(
+        name="blacklist", description="Blacklist a user from giveaway"
     )
-    async def add_giveaway_blacklist(
+    @app_commands.describe(
+        big="Yes if the giveaway is more than Rp50.000",
+        winners="Discord IDs of the giveaway winner. Example: 1234567890,0987654321",
+        ghosts="Member IDs who do not claim the giveaway. Example: 1234567890,0987654321",
+    )
+    async def add_giveaway_blacklist(  # noqa: C901, PLR0912, FIX002 # TODO: Improve this function
         self,
         interaction: Interaction,
         big: bool,
         winners: str,
-        ghosts: Optional[str] = None,
+        ghosts: str | None = None,
     ) -> None:
         await interaction.response.defer()
+        if not interaction.guild:
+            return None
+
         if not interaction.user.guild_permissions.administrator:
-            return await interaction.followup.send('You are not an admin', ephemeral=True)
+            return await interaction.followup.send(
+                "You are not an admin", ephemeral=True
+            )
 
         winner_list: list[discord.Member] = []
-        winners = winners.split(',')
-        for winner in winners:
+        winners_list = winners.split(",")
+        for winner in winners_list:
             if (user := interaction.guild.get_member(int(winner))) is None:
-                return await interaction.followup.send(f'User {winner} not found', ephemeral=True)
+                return await interaction.followup.send(
+                    f"User {winner} not found", ephemeral=True
+                )
             winner_list.append(user)
 
         ghost_list: list[discord.Member] = []
         if ghosts:
-            ghosts = ghosts.split(',')
-            for ghost in ghosts:
+            ghosts_list = ghosts.split(",")
+            for ghost in ghosts_list:
                 if (user := interaction.guild.get_member(int(ghost))) is None:
                     return await interaction.followup.send(
-                        f'User {ghost} not found', ephemeral=True
+                        f"User {ghost} not found", ephemeral=True
                     )
                 ghost_list.append(user)
 
@@ -64,27 +80,33 @@ class Giveaway(commands.GroupCog, group_name='warnet-ga'):
             winner_day = 15
             ghosting_day = 7
 
-        now = datetime.now(timezone.utc)
+        now = datetime.now(UTC)
         end_time_no_streak = now + timedelta(days=winner_day)
         end_time_streak = now + timedelta(days=winner_day * 2)
         end_time_ghosting = now + timedelta(days=ghosting_day)
         blacklist_role = interaction.guild.get_role(BLACKLIST_GA_ROLE_ID)
+        if blacklist_role is None:
+            logger.error(
+                "Role not found",
+                extra={"role_id": BLACKLIST_GA_ROLE_ID},
+            )
+            return None
 
         async with self.db_pool.acquire() as conn:
             streak_user_records = await conn.fetch(
-                'SELECT user_id FROM blacklist_ga WHERE status_user = 1',
+                "SELECT user_id FROM blacklist_ga WHERE status_user = 1",
             )
-            streak_user = {record['user_id'] for record in streak_user_records}
+            streak_user = {record["user_id"] for record in streak_user_records}
 
             for winner in winner_list:
                 await winner.add_roles(blacklist_role)
                 if winner.id in streak_user:
                     await conn.execute(
-                        '''
-                        UPDATE blacklist_ga 
+                        """
+                        UPDATE blacklist_ga
                         SET end_time = $2, status_user = $3, cooldown_time = $4
                         WHERE user_id = $1
-                        ''',
+                        """,
                         winner.id,
                         end_time_streak,
                         0,
@@ -92,33 +114,43 @@ class Giveaway(commands.GroupCog, group_name='warnet-ga'):
                     )
                 else:
                     await conn.execute(
-                        'INSERT INTO blacklist_ga (user_id, end_time, status_user, cooldown_time) VALUES ($1, $2, $3, $4)',
+                        "INSERT INTO blacklist_ga (user_id, end_time, status_user, cooldown_time) VALUES ($1, $2, $3, $4)",
                         winner.id,
                         end_time_no_streak,
                         1,
                         end_time_streak,
                     )
                 logger.info(
-                    f'Added role {blacklist_role.id} to user {winner.id} for {winner_day} days'
+                    "Added role blacklist role to winner",
+                    extra={
+                        "role_id": blacklist_role.id,
+                        "winner_id": winner.id,
+                        "winned_day": winner_day,
+                    },
                 )
 
             for ghost in ghost_list:
                 await ghost.add_roles(blacklist_role)
                 if ghost.id in streak_user:
                     await conn.execute(
-                        'UPDATE blacklist_ga SET end_time = $2 WHERE user_id = $1',
+                        "UPDATE blacklist_ga SET end_time = $2 WHERE user_id = $1",
                         ghost.id,
                         end_time_ghosting,
                     )
                 else:
                     await conn.execute(
-                        'INSERT INTO blacklist_ga (user_id, end_time, status_user) VALUES ($1, $2, $3)',
+                        "INSERT INTO blacklist_ga (user_id, end_time, status_user) VALUES ($1, $2, $3)",
                         ghost.id,
                         end_time_ghosting,
                         0,
                     )
                 logger.info(
-                    f'Added role {blacklist_role.id} to user {ghost.id} for {ghosting_day} days'
+                    "Added role blacklist role to ghost",
+                    extra={
+                        "role_id": blacklist_role.id,
+                        "winner_id": ghost.id,
+                        "winned_day": winner_day,
+                    },
                 )
 
         embed = discord.Embed(
@@ -134,42 +166,53 @@ class Giveaway(commands.GroupCog, group_name='warnet-ga'):
             name="List of Ghosts",
             value=f"{', '.join([ghost.mention for ghost in ghost_list])}",
         )
-        await interaction.followup.send(embed=embed)
+        return await interaction.followup.send(embed=embed)
 
     @tasks.loop(seconds=60)
     async def _check_blacklist_ga(self) -> None:
         guild = self.bot.get_guild(GUILD_ID)
+        if guild is None:
+            logger.error("Guild not found")
+            return
+
         blacklist_role = guild.get_role(BLACKLIST_GA_ROLE_ID)
-        now = datetime.now(timezone.utc)
+        if blacklist_role is None:
+            logger.error("Blacklist role not found")
+            return
+
+        now = datetime.now(UTC)
 
         async with self.db_pool.acquire() as conn:
             user_want_remove = await conn.fetch(
-                'SELECT user_id FROM blacklist_ga WHERE end_time <= $1 AND has_role = TRUE',
+                "SELECT user_id FROM blacklist_ga WHERE end_time <= $1 AND has_role = TRUE",
                 now,
             )
             for user in user_want_remove:
-                if user := guild.get_member(int(user['user_id'])):
+                if user := guild.get_member(int(user["user_id"])):
                     await user.remove_roles(blacklist_role)
                     await conn.execute(
-                        'UPDATE blacklist_ga SET has_role = FALSE WHERE user_id = $1',
+                        "UPDATE blacklist_ga SET has_role = FALSE WHERE user_id = $1",
                         user.id,
                     )
-                    logger.info(f'Removed role {blacklist_role.id} from user {user.id} (rm role)')
+                    logger.info(
+                        "Removed role  from user (rm role)",
+                        extra={"role": blacklist_role.id, "user": user.id},
+                    )
 
             user_want_delete = await conn.fetch(
-                'SELECT user_id FROM blacklist_ga WHERE has_role = FALSE AND status_user = 0 OR cooldown_time <= $1',
+                "SELECT user_id FROM blacklist_ga WHERE has_role = FALSE AND status_user = 0 OR cooldown_time <= $1",
                 now,
             )
             for user in user_want_delete:
                 if user:
                     await conn.execute(
-                        'DELETE FROM blacklist_ga WHERE user_id = $1',
-                        int(user['user_id']),
+                        "DELETE FROM blacklist_ga WHERE user_id = $1",
+                        int(user["user_id"]),
                     )
-                    logger.info(f'Deleted user {user} (rm row table)')
+                    logger.info("Deleted user (rm row table)", extra=user)
 
     @_check_blacklist_ga.before_loop
-    async def _before_check_blacklist_ga(self):
+    async def _before_check_blacklist_ga(self) -> None:
         await self.bot.wait_until_ready()
